some improvements
Making files easier to navigate
Added restart command
