+++ DynaBA/Models/BotModel.cs
@@ -3,4 +3,5 @@
 public struct Bot
 {
     public string Token { get; set; }
+    public string? RebootMsg { get; set; }
 }
